requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-redis==5.4.0
sentry-sdk
djangorestframework-simplejwt